refactor: split clash into library + thin CLI binary
Extract core logic from main.rs into a reusable library so clash can be
used as a dependency. The clash crate now builds both a library (lib.rs)
and a binary (main.rs).

New modules:
- lib.rs: public API re-exporting all modules
- handlers.rs: pre-built hook handlers (handle_permission_request,
  handle_session_start, etc.) moved from main.rs

Moved to settings.rs:
- DEFAULT_POLICY constant
- ClashSettings::set_policy() for programmatic construction

main.rs is now a thin CLI wrapper that imports from the library.

All 281 tests pass, no new warnings.
